refactor: 💡 move all JobRunnerError to libcommon
Two ideas behind the refactoring:
1. having libcommon aware of the list of error codes, to be able to
   relate with list of error codes to retry, in a next PR
2. centralize all the error codes, since they share the same namespace.
   It will help do support.
