+++ services/worker/src/worker/job_runners/config/info.py
@@ -1,37 +1,11 @@
 import logging
-from http import HTTPStatus
-from typing import Any, Dict, Literal, Optional, TypedDict
+from typing import Any, Dict, TypedDict
 
 from libcommon.constants import PROCESSING_STEP_CONFIG_INFO_VERSION
+from libcommon.exceptions import PreviousStepFormatError
 
-from worker.common_exceptions import JobRunnerError
 from worker.job_runners.config.config_job_runner import ConfigJobRunner
 from worker.utils import CompleteJobResult, get_previous_step_or_raise
-
-ConfigInfoJobRunnerErrorCode = Literal["PreviousStepFormatError"]
-
-
-class ConfigInfoJobRunnerError(JobRunnerError):
-    """Base class for exceptions in this module."""
-
-    def __init__(
-        self,
-        message: str,
-        status_code: HTTPStatus,
-        code: ConfigInfoJobRunnerErrorCode,
-        cause: Optional[BaseException] = None,
-        disclose_cause: bool = False,
-    ):
-        super().__init__(
-            message=message, status_code=status_code, code=code, cause=cause, disclose_cause=disclose_cause
-        )
-
-
-class PreviousStepFormatError(ConfigInfoJobRunnerError):
-    """Raised when the content of the previous step has not the expected format."""
-
-    def __init__(self, message: str, cause: Optional[BaseException] = None):
-        super().__init__(message, HTTPStatus.INTERNAL_SERVER_ERROR, "PreviousStepFormatError", cause, False)
 
 
 class ConfigInfoResponse(TypedDict):
