+++ services/worker/src/worker/job_runners/config/parquet.py
@@ -2,44 +2,18 @@
 # Copyright 2022 The HuggingFace Authors.
 
 import logging
-from http import HTTPStatus
-from typing import List, Literal, Optional, TypedDict
+from typing import List, TypedDict
 
 from libcommon.constants import PROCESSING_STEP_CONFIG_PARQUET_VERSION
+from libcommon.exceptions import PreviousStepFormatError
 
-from worker.common_exceptions import JobRunnerError
 from worker.job_runners.config.config_job_runner import ConfigJobRunner
 from worker.job_runners.config.parquet_and_info import ParquetFileItem
 from worker.utils import CompleteJobResult, get_previous_step_or_raise
 
-ConfigParquetJobRunnerErrorCode = Literal["PreviousStepFormatError"]
-
 
 class ConfigParquetResponse(TypedDict):
     parquet_files: List[ParquetFileItem]
-
-
-class ConfigParquetJobRunnerError(JobRunnerError):
-    """Base class for exceptions in this module."""
-
-    def __init__(
-        self,
-        message: str,
-        status_code: HTTPStatus,
-        code: ConfigParquetJobRunnerErrorCode,
-        cause: Optional[BaseException] = None,
-        disclose_cause: bool = False,
-    ):
-        super().__init__(
-            message=message, status_code=status_code, code=code, cause=cause, disclose_cause=disclose_cause
-        )
-
-
-class PreviousStepFormatError(ConfigParquetJobRunnerError):
-    """Raised when the content of the previous step has not the expected format."""
-
-    def __init__(self, message: str, cause: Optional[BaseException] = None):
-        super().__init__(message, HTTPStatus.INTERNAL_SERVER_ERROR, "PreviousStepFormatError", cause, False)
 
 
 def compute_parquet_response(dataset: str, config: str) -> ConfigParquetResponse:
