+++ services/worker/src/worker/job_runners/config/size.py
@@ -2,16 +2,13 @@
 # Copyright 2022 The HuggingFace Authors.
 
 import logging
-from http import HTTPStatus
-from typing import Literal, Optional, TypedDict
+from typing import TypedDict
 
 from libcommon.constants import PROCESSING_STEP_CONFIG_SIZE_VERSION
+from libcommon.exceptions import PreviousStepFormatError
 
-from worker.common_exceptions import JobRunnerError
 from worker.job_runners.config.config_job_runner import ConfigJobRunner
 from worker.utils import CompleteJobResult, get_previous_step_or_raise
-
-ConfigSizeJobRunnerErrorCode = Literal["PreviousStepFormatError"]
 
 
 class ConfigSize(TypedDict):
@@ -41,29 +38,6 @@ class ConfigSizeContent(TypedDict):
 
 class ConfigSizeResponse(TypedDict):
     size: ConfigSizeContent
-
-
-class ConfigSizeJobRunnerError(JobRunnerError):
-    """Base class for exceptions in this module."""
-
-    def __init__(
-        self,
-        message: str,
-        status_code: HTTPStatus,
-        code: ConfigSizeJobRunnerErrorCode,
-        cause: Optional[BaseException] = None,
-        disclose_cause: bool = False,
-    ):
-        super().__init__(
-            message=message, status_code=status_code, code=code, cause=cause, disclose_cause=disclose_cause
-        )
-
-
-class PreviousStepFormatError(ConfigSizeJobRunnerError):
-    """Raised when the content of the previous step has not the expected format."""
-
-    def __init__(self, message: str, cause: Optional[BaseException] = None):
-        super().__init__(message, HTTPStatus.INTERNAL_SERVER_ERROR, "PreviousStepFormatError", cause, False)
 
 
 def compute_config_size_response(dataset: str, config: str) -> ConfigSizeResponse:
