+++ services/worker/src/worker/job_runners/config/split_names_from_streaming.py
@@ -2,8 +2,7 @@
 # Copyright 2022 The HuggingFace Authors.
 
 import logging
-from http import HTTPStatus
-from typing import List, Literal, Optional, Union
+from typing import List, Optional, Union
 
 from datasets import get_dataset_split_names
 from datasets.data_files import EmptyDatasetError as _EmptyDatasetError
@@ -11,45 +10,10 @@ from libcommon.constants import (
     PROCESSING_STEP_SPLIT_NAMES_FROM_DATASET_INFO_VERSION,
     PROCESSING_STEP_SPLIT_NAMES_FROM_STREAMING_VERSION,
 )
+from libcommon.exceptions import EmptyDatasetError, SplitNamesFromStreamingError
 
-from worker.common_exceptions import JobRunnerError
 from worker.job_runners.config.config_job_runner import ConfigCachedJobRunner
 from worker.utils import CompleteJobResult, JobRunnerInfo, SplitItem, SplitsList
-
-SplitNamesFromStreamingJobRunnerErrorCode = Literal[
-    "EmptyDatasetError",
-    "SplitNamesFromStreamingError",
-]
-
-
-class SplitNamesFromStreamingJobRunnerError(JobRunnerError):
-    """Base class for split names job runner exceptions."""
-
-    def __init__(
-        self,
-        message: str,
-        status_code: HTTPStatus,
-        code: SplitNamesFromStreamingJobRunnerErrorCode,
-        cause: Optional[BaseException] = None,
-        disclose_cause: bool = False,
-    ):
-        super().__init__(
-            message=message, status_code=status_code, code=code, cause=cause, disclose_cause=disclose_cause
-        )
-
-
-class SplitNamesFromStreamingError(SplitNamesFromStreamingJobRunnerError):
-    """Raised when the split names could not be fetched."""
-
-    def __init__(self, message: str, cause: Optional[BaseException] = None):
-        super().__init__(message, HTTPStatus.INTERNAL_SERVER_ERROR, "SplitNamesFromStreamingError", cause, True)
-
-
-class EmptyDatasetError(SplitNamesFromStreamingJobRunnerError):
-    """Raised when the dataset has no data."""
-
-    def __init__(self, message: str, cause: Optional[BaseException] = None):
-        super().__init__(message, HTTPStatus.INTERNAL_SERVER_ERROR, "EmptyDatasetError", cause, True)
 
 
 def compute_split_names_from_streaming_response(
