+++ services/worker/src/worker/job_runners/dataset/size.py
@@ -3,17 +3,15 @@
 
 import logging
 from http import HTTPStatus
-from typing import Literal, Optional, Tuple, TypedDict
+from typing import Tuple, TypedDict
 
 from libcommon.constants import PROCESSING_STEP_DATASET_SIZE_VERSION
+from libcommon.exceptions import PreviousStepFormatError
 from libcommon.simple_cache import DoesNotExist, get_response
 
-from worker.common_exceptions import JobRunnerError
 from worker.job_runners.config.size import ConfigSize, ConfigSizeResponse, SplitSize
 from worker.job_runners.dataset.dataset_job_runner import DatasetJobRunner
 from worker.utils import JobResult, PreviousJob, get_previous_step_or_raise
-
-SizesJobRunnerErrorCode = Literal["PreviousStepFormatError"]
 
 
 class DatasetSize(TypedDict):
@@ -34,29 +32,6 @@ class DatasetSizeResponse(TypedDict):
     size: DatasetSizeContent
     pending: list[PreviousJob]
     failed: list[PreviousJob]
-
-
-class DatasetSizeJobRunnerError(JobRunnerError):
-    """Base class for exceptions in this module."""
-
-    def __init__(
-        self,
-        message: str,
-        status_code: HTTPStatus,
-        code: SizesJobRunnerErrorCode,
-        cause: Optional[BaseException] = None,
-        disclose_cause: bool = False,
-    ):
-        super().__init__(
-            message=message, status_code=status_code, code=code, cause=cause, disclose_cause=disclose_cause
-        )
-
-
-class PreviousStepFormatError(DatasetSizeJobRunnerError):
-    """Raised when the content of the previous step has not the expected format."""
-
-    def __init__(self, message: str, cause: Optional[BaseException] = None):
-        super().__init__(message, HTTPStatus.INTERNAL_SERVER_ERROR, "PreviousStepFormatError", cause, False)
 
 
 def compute_sizes_response(dataset: str) -> Tuple[DatasetSizeResponse, float]:
