+++ services/worker/src/worker/job_runners/split/split_job_runner.py
@@ -3,10 +3,10 @@
 
 from pathlib import Path
 
+from libcommon.exceptions import ParameterMissingError
 from libcommon.processing_graph import ProcessingStep
 from libcommon.utils import JobInfo
 
-from worker.common_exceptions import ParameterMissingError
 from worker.config import AppConfig
 from worker.job_runners._datasets_based_job_runner import DatasetsBasedJobRunner
 from worker.job_runners.config.config_job_runner import ConfigJobRunner
