+++ services/worker/tests/job_runners/config/test_info.py
@@ -5,14 +5,14 @@ from http import HTTPStatus
 from typing import Any, Callable
 
 import pytest
+from libcommon.exceptions import PreviousStepError, PreviousStepFormatError
 from libcommon.processing_graph import ProcessingGraph
 from libcommon.resources import CacheMongoResource, QueueMongoResource
 from libcommon.simple_cache import upsert_response
 from libcommon.utils import Priority
 
-from worker.common_exceptions import PreviousStepError
 from worker.config import AppConfig
-from worker.job_runners.config.info import ConfigInfoJobRunner, PreviousStepFormatError
+from worker.job_runners.config.info import ConfigInfoJobRunner
 
 
 @pytest.fixture(autouse=True)
