+++ services/worker/tests/job_runners/config/test_parquet.py
@@ -5,17 +5,16 @@ from http import HTTPStatus
 from typing import Any, Callable
 
 import pytest
+from libcommon.exceptions import PreviousStepError, PreviousStepFormatError
 from libcommon.processing_graph import ProcessingGraph
 from libcommon.resources import CacheMongoResource, QueueMongoResource
 from libcommon.simple_cache import upsert_response
 from libcommon.utils import Priority
 
-from worker.common_exceptions import PreviousStepError
 from worker.config import AppConfig
 from worker.job_runners.config.parquet import (
     ConfigParquetJobRunner,
     ConfigParquetResponse,
-    PreviousStepFormatError,
 )
 from worker.job_runners.config.parquet_and_info import (
     ConfigParquetAndInfoResponse,
