+++ services/worker/tests/job_runners/config/test_parquet_and_info.py
@@ -14,7 +14,14 @@ import pytest
 import requests
 from datasets import Audio, Features, Image, Value
 from huggingface_hub.hf_api import HfApi
-from libcommon.exceptions import CustomError
+from libcommon.exceptions import (
+    CustomError,
+    DatasetInBlockListError,
+    DatasetTooBigFromDatasetsError,
+    DatasetTooBigFromHubError,
+    DatasetWithTooBigExternalFilesError,
+    DatasetWithTooManyExternalFilesError,
+)
 from libcommon.processing_graph import ProcessingGraph
 from libcommon.resources import CacheMongoResource, QueueMongoResource
 from libcommon.simple_cache import upsert_response
@@ -23,11 +30,6 @@ from libcommon.utils import Priority
 from worker.config import AppConfig
 from worker.job_runners.config.parquet_and_info import (
     ConfigParquetAndInfoJobRunner,
-    DatasetInBlockListError,
-    DatasetTooBigFromDatasetsError,
-    DatasetTooBigFromHubError,
-    DatasetWithTooBigExternalFilesError,
-    DatasetWithTooManyExternalFilesError,
     get_dataset_info_or_raise,
     get_writer_batch_size,
     parse_repo_filename,
