+++ services/worker/tests/job_runners/config/test_split_names_from_dataset_info.py
@@ -6,16 +6,14 @@ from typing import Any, Callable
 from unittest.mock import Mock
 
 import pytest
-from libcommon.exceptions import CustomError
+from libcommon.exceptions import CustomError, PreviousStepError, PreviousStepFormatError
 from libcommon.processing_graph import ProcessingGraph
 from libcommon.resources import CacheMongoResource, QueueMongoResource
 from libcommon.simple_cache import upsert_response
 from libcommon.utils import Priority
 
-from worker.common_exceptions import PreviousStepError
 from worker.config import AppConfig
 from worker.job_runners.config.split_names_from_dataset_info import (
-    PreviousStepFormatError,
     SplitNamesFromDatasetInfoJobRunner,
 )
