+++ services/worker/tests/job_runners/dataset/test_opt_in_out_urls_count.py
@@ -5,12 +5,12 @@ from http import HTTPStatus
 from typing import Any, Callable, List
 
 import pytest
+from libcommon.exceptions import PreviousStepError
 from libcommon.processing_graph import ProcessingGraph
 from libcommon.resources import CacheMongoResource, QueueMongoResource
 from libcommon.simple_cache import upsert_response
 from libcommon.utils import Priority
 
-from worker.common_exceptions import PreviousStepError
 from worker.config import AppConfig
 from worker.job_runners.dataset.opt_in_out_urls_count import (
     DatasetOptInOutUrlsCountJobRunner,
