+++ services/worker/tests/job_runners/dataset/test_parquet.py
@@ -5,19 +5,18 @@ from http import HTTPStatus
 from typing import Any, Callable
 
 import pytest
+from libcommon.exceptions import PreviousStepError, PreviousStepFormatError
 from libcommon.processing_graph import ProcessingGraph
 from libcommon.resources import CacheMongoResource, QueueMongoResource
 from libcommon.simple_cache import upsert_response
 from libcommon.utils import Priority
 
-from worker.common_exceptions import PreviousStepError
 from worker.config import AppConfig
 from worker.job_runners.config.parquet import ConfigParquetResponse
 from worker.job_runners.config.parquet_and_info import ParquetFileItem
 from worker.job_runners.dataset.parquet import (
     DatasetParquetJobRunner,
     DatasetParquetResponse,
-    PreviousStepFormatError,
 )
 
 from ..utils import UpstreamResponse
