+++ services/worker/tests/job_runners/dataset/test_size.py
@@ -5,17 +5,14 @@ from http import HTTPStatus
 from typing import Any, Callable
 
 import pytest
+from libcommon.exceptions import PreviousStepError, PreviousStepFormatError
 from libcommon.processing_graph import ProcessingGraph
 from libcommon.resources import CacheMongoResource, QueueMongoResource
 from libcommon.simple_cache import upsert_response
 from libcommon.utils import Priority
 
-from worker.common_exceptions import PreviousStepError
 from worker.config import AppConfig
-from worker.job_runners.dataset.size import (
-    DatasetSizeJobRunner,
-    PreviousStepFormatError,
-)
+from worker.job_runners.dataset.size import DatasetSizeJobRunner
 
 from ..utils import UpstreamResponse
